refactor items stuff out of index
