+++ src/index.js
@@ -1,4 +1,3 @@
-import * as R from "ramda";
 import * as dotenv from "dotenv";
 
 import { App, MemoryStore } from "@slack/bolt";
@@ -23,16 +22,13 @@ import {
   viewProject
 } from "./projects";
 
+import { handleItemMod, saveItem } from "./items";
+
 import {
-  addItem,
   addSection,
-  deleteItem,
   deleteSection,
-  getItemById,
   getSectionById,
-  moveItem,
   moveSection,
-  updateItem,
   updateSection
 } from "./models";
 
@@ -96,66 +92,7 @@ app.command("/alfred", async ({ command, ack, respond, context, body }) => {
   }
 });
 
-app.view(ACTIONS.saveItem, async ({ ack, body, view, context }) => {
-  ack();
-  const { values } = view.state;
-  const itemName = R.pathOr("", ["item_name", "item_name", "value"], values);
-  const itemUrl =
-    R.path(["item_url", "item_url", "value"], values) ||
-    R.path(["item_user", "item_user", "selected_user"], values) ||
-    R.path(["item_channel", "item_channel", "selected_channel"], values);
-  const itemDescription = R.pathOr(
-    "",
-    ["item_description", "item_description", "value"],
-    values
-  );
-
-  const { id, sectionId, type } = JSON.parse(view.private_metadata);
-  const itemId = id;
-
-  const convo = convoStore.get(body.user.id);
-  if (itemId) {
-    try {
-      await updateItem(itemId, itemName, itemUrl, itemDescription, type);
-      convo.then(async ({ respond, token, projectName }) => {
-        postAuditMessage(
-          body.user.id,
-          projectName,
-          `${itemName}: ${itemUrl}, ${itemDescription}`,
-          context.botToken
-        );
-        await lookupProject(projectName, true, respond, token);
-      });
-    } catch (err) {
-      console.log("error in ACTIONS.saveItem (updateItem)", err);
-      app.client.chat.postMessage({
-        token: context.botToken,
-        channel: body.user.id,
-        text: MESSAGES.genericError("update that item")
-      });
-    }
-  } else {
-    try {
-      await addItem(itemName, sectionId, itemUrl, itemDescription, type);
-      convo.then(async ({ respond, token, projectName }) => {
-        postAuditMessage(
-          body.user.id,
-          projectName,
-          `${itemName}: ${itemUrl}, ${itemDescription}`,
-          context.botToken
-        );
-        await lookupProject(projectName, true, respond, token);
-      });
-    } catch (err) {
-      console.log("error in ACTIONS.saveItem (addItem)", err);
-      app.client.chat.postMessage({
-        token: context.botToken,
-        channel: body.user.id,
-        text: MESSAGES.genericError("add that new item")
-      });
-    }
-  }
-});
+app.view(ACTIONS.saveItem, saveItem(app, convoStore));
 
 app.view(ACTIONS.saveSection, async ({ ack, body, view, context }) => {
   ack();
@@ -297,65 +234,4 @@ app.action(
   }
 );
 
-app.action(ACTIONS.modItem, async ({ action, ack, context, body, respond }) => {
-  ack();
-  const actionValue = JSON.parse(action.selected_option.value);
-  const command = actionValue.cmd;
-  const projectName = actionValue.pn;
-  const itemId = actionValue.iId;
-
-  switch (command) {
-    case COMMANDS.edit:
-      convoStore.set(body.user.id, {
-        respond,
-        token: context.botToken,
-        projectName
-      });
-      try {
-        const item = await getItemById(itemId);
-        const blocks = itemModal(item);
-        app.client.views.open({
-          token: context.botToken,
-          view: blocks,
-          trigger_id: body.trigger_id
-        });
-      } catch (err) {
-        console.log("error in ACTIONS.modItem (edit)", err);
-        app.client.chat.postMessage({
-          token: context.botToken,
-          channel: body.user.id,
-          text: MESSAGES.genericError("edit that item")
-        });
-      }
-      break;
-    case COMMANDS.up:
-    case COMMANDS.down:
-      try {
-        await moveItem(itemId, command);
-        await lookupProject(projectName, true, respond, context.botToken);
-      } catch (err) {
-        console.log("error in ACTIONS.modItem (move)", err);
-        respond({
-          token: context.botToken,
-          response_type: "ephemeral",
-          text: MESSAGES.genericError("move that")
-        });
-      }
-      break;
-    case COMMANDS.delete:
-      try {
-        await deleteItem(itemId);
-        await lookupProject(projectName, true, respond, context.botToken);
-      } catch (err) {
-        console.log("error in ACTIONS.modItem (delete)", err);
-        respond({
-          token: context.botToken,
-          response_type: "ephemeral",
-          text: MESSAGES.genericError("delete that")
-        });
-      }
-      break;
-    default:
-      console.log("Shouldn't be able to do this...");
-  }
-});
+app.action(ACTIONS.modItem, handleItemMod(app, convoStore));
